// FirstApp/src/generated/models/Cr989_ar_codeappsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr989_ar_codeappscr989_typeofaccess = {
  727490000: "Test",
  727490001: "Shadow",
  727490002: "Business",
  727490003: "Legal",
  727490004: "Other",
} as const;
export type Cr989_ar_codeappscr989_typeofaccess =
  keyof typeof Cr989_ar_codeappscr989_typeofaccess;
export const Cr989_ar_codeappsstatecode = {
  0: "Active",
  1: "Inactive",
} as const;
export type Cr989_ar_codeappsstatecode =
  keyof typeof Cr989_ar_codeappsstatecode;
export const Cr989_ar_codeappsstatuscode = {
  1: "Active",
  2: "Inactive",
} as const;
export type Cr989_ar_codeappsstatuscode =
  keyof typeof Cr989_ar_codeappsstatuscode;

export interface Cr989_ar_codeappsBase {
  cr989_accesscategory?: string;
  cr989_accessrightname?: string;
  cr989_ar_codeappid: string;
  cr989_businessunit?: string;
  cr989_createdon?: string;
  cr989_customername?: string;
  cr989_customflag?: string;
  cr989_customgrouping?: string;
  cr989_gatekeeperaccessright?: string;
  cr989_isautogranted?: boolean;
  cr989_legalentity?: string;
  cr989_levelcode?: string;
  cr989_modifiedon?: string;
  cr989_offering?: string;
  cr989_onbehalfof?: string;
  cr989_ownername?: string;
  cr989_requestedaccessright?: string;
  cr989_roleaccesstype?: string;
  cr989_rolename?: string;
  cr989_roletype?: string;
  cr989_statusreason?: string;
  cr989_typeofaccess?: Cr989_ar_codeappscr989_typeofaccess;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr989_ar_codeappsstatecode;
  statuscode?: Cr989_ar_codeappsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr989_ar_codeapps extends Cr989_ar_codeappsBase {
  cr989_isautograntedname?: string;
  cr989_typeofaccessname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}